Various changes. Note long commit message follows.
(1) Add options
    --tag-regex
and
    --branch-regex

to allow multiple tags or multiple branches to be fetched.

Example:

   esgfetchtables --project cmip6 --branch-regex '[0-9\.]+$'

(2) Store the checksum and download URL for each file from the initial API
response relating to the whole directory (not only the name), avoiding the
need for API call for file info for each file individually.  This makes it
much faster, and far fewer API calls.  There is also a single API call for
the CMIP6_CV.json file done before any looping over branches/tags, because
this is always from master.

(3) The "ctx.error" property is changed from a Boolean to a string or None,
so that the status can be properly displayed in case of an exception.